Change of response on calculation of a tax.

+++ src/main/java/ar/edu/unq/ttip/alec/backend/model/tax/TaxStrategy.java
@@ -2,7 +2,9 @@ package ar.edu.unq.ttip.alec.backend.model.tax;
 
 import ar.edu.unq.ttip.alec.backend.model.Apartado;
 
+import java.math.BigDecimal;
+
 public interface TaxStrategy {
 
-    Double calculateWith(Double amount, Apartado apartado);
+    BigDecimal calculateWith(BigDecimal amount, Apartado apartado);
 }
