+++ src/main/java/ar/edu/unq/ttip/alec/backend/service/dtos/CalculationDTO.java
@@ -2,13 +2,15 @@ package ar.edu.unq.ttip.alec.backend.service.dtos;
 
 import ar.edu.unq.ttip.alec.backend.model.Apartado;
 
+import java.math.BigDecimal;
+
 public class CalculationDTO {
 
-    private Double amount;
+    private BigDecimal amount;
     private Apartado apartado;
     private Integer taxId;
 
-    public CalculationDTO(Double amount, Apartado apartado, Integer taxId){
+    public CalculationDTO(BigDecimal amount, Apartado apartado, Integer taxId){
         this.amount=amount;
         this.apartado=apartado;
         this.taxId=taxId;
@@ -16,7 +18,7 @@ public class CalculationDTO {
 
     protected CalculationDTO(){}
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
